Rename function to IsInsideExternalXtfResource
