+++ src/test/data/IsInsideExternalXtfResource/MandatoryConstraintThis.ili
@@ -1,0 +1,32 @@
+INTERLIS 2.4;
+
+MODEL TestSuite
+  AT "mailto:info@geowerkstatt.ch" VERSION "2023-12-20" =
+  IMPORTS GeoW_FunctionsExt;
+
+  DOMAIN
+      !!@CRS=EPSG:2056
+      CHKoord = COORD 2460000.000 .. 2870000.000 [INTERLIS.m],
+                      1045000.000 .. 1310000.000 [INTERLIS.m],
+                      ROTATION 2 -> 1;
+
+  TOPIC FunctionTestTopic =
+    CLASS ConstraintTestClass (ABSTRACT) =
+        expected : MANDATORY BOOLEAN;
+    END ConstraintTestClass;
+
+    CLASS IsInsideSOKantonsgrenze EXTENDS ConstraintTestClass =
+        testAttributeIsInsideSOKantonsgrenze : CHKoord;
+        MANDATORY CONSTRAINT IsInsideSOKantonsgrenze: GeoW_FunctionsExt.IsInsideExternalXtfResource("IsInsideExternalXtfResource/ch.so.agi.av.hoheitsgrenzen_excerpt.xtf", "SO_Hoheitsgrenzen_Publikation_20170626.Hoheitsgrenzen.Kantonsgrenze.Geometrie", "fubar", THIS, "testAttributeIsInsideSOKantonsgrenze") == expected;
+    END IsInsideSOKantonsgrenze;
+
+    CLASS InvalidConstraints =
+        geometryAttribute : CHKoord;
+        MANDATORY CONSTRAINT MalformedDatasetName: GeoW_FunctionsExt.IsInsideExternalXtfResource("IsInsideExternalXtfResource/ch.so.agi.av.hoheitsgrenzen_excerpt.xtf", "DatasetNameWithoutQualifiedAttribute", "fubar", THIS, "geometryAttribute");
+        MANDATORY CONSTRAINT NonExistentTransferFile: GeoW_FunctionsExt.IsInsideExternalXtfResource("NotExistingFile.xtf", "SO_Hoheitsgrenzen_Publikation_20170626.Hoheitsgrenzen.Kantonsgrenze.Geometrie", "fubar", THIS, "geometryAttribute");
+        MANDATORY CONSTRAINT NonExistentTransferIds: GeoW_FunctionsExt.IsInsideExternalXtfResource("IsInsideExternalXtfResource/ch.so.agi.av.hoheitsgrenzen_excerpt.xtf", "SO_Hoheitsgrenzen_Publikation_20170626.Hoheitsgrenzen.Kantonsgrenze.Geometrie", "9999,100000000", THIS, "geometryAttribute");
+    END InvalidConstraints;
+
+  END FunctionTestTopic;
+
+END TestSuite.
