+++ src/test/data/IsInsideExternalXtfResource/TestData.xtf
