+++ src/test/resources/IsInsideExternalXtfResource/ch.so.agi.av.hoheitsgrenzen_excerpt.xtf
